feat(server): snapshot restore carries createRequestId into pane create bodies
